Bugfix Movement animation
A minor Bugfix, which caused the Movement animation to not play correclty. If Player wanted to move downwards, the the animation for moving upward has been played.

Reason was the Movement script in Player, which sets the float from the animator "Vertical", but in the script it was written "Vertival".